新的复用GPIO模块，更新固件库

+++ firmware_lib/rust/examples/key_sw.rs
@@ -29,6 +29,6 @@ fn main() -> ! {
     }
 }
 
-const fn bit_is_1(num: u16, bit: u8) -> bool {
+const fn bit_is_1(num: u8, bit: u8) -> bool {
     (num & (1 << bit)) != 0
 }
